Refactor processLiquidityRewardEarnedNotification
Refactor processLiquidityRewardEarnedNotification function and get rid of
unnecessary chunks of code.
